Starting to fix execution problems
