+++ DungeonProject/DungeonProject/Utilities.py
@@ -1,4 +1,5 @@
 import sys
+import random
 
 class Utilities:
 
@@ -19,6 +20,15 @@ class Utilities:
                 graphs[-1] += c
                 depth += 1 if c == "[" else -1 if c == "]" else 0
         return graphs
+
+    # Returns a random key using the given weights
+    # Data must be inputed as a dictionary, with choices as keys and weights as values.
+    @staticmethod
+    def randomWeightedChoice(data):
+        cumulative = [sum(data.values()[:i+1]) for i in range(len(data.values()))]
+        value = random.uniform(0, cumulative[-1])
+        index = next(i for i in range(len(cumulative)) if cumulative[i] > value)
+        return data.keys()[index]
